backend-app/requirements.txt
fastapi
uvicorn
boto3
pytesseract
gTTS
firebase-admin
translate
python-dotenv
opencv-python-headless